factor partials into separate module
